requirements.txt
# Core
PyYAML>=6.0.1

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Development (optional)
# black>=24.0.0
# isort>=5.13.0
# ruff>=0.1.0
# pytest-cov>=4.1.0
